refactor(auth): 에러 폴더 구조와 DTO 경계 정리
