+++ src/clips/application/usecases/list-favorite-clips.usecase.ts
@@ -6,7 +6,7 @@ import {
   ClipTypeFilter,
 } from '../../domain/clips.repository';
 import type { ClipsRepository } from '../../domain/clips.repository';
-import { ClipsError } from '../clips.error';
+import { ClipsError } from '../errors/clips.error';
 import {
   LIST_CLIPS_LIMIT,
   buildPage,
