--- src/folders/dtos/create-folder.dto.ts
@@ -1,7 +1,0 @@
-import { IsNotEmpty, IsString } from 'class-validator';
-
-export class CreateFolderDto {
-  @IsString()
-  @IsNotEmpty()
-  name: string;
-}
