--- src/folders/dtos/reorder-folder.dto.ts
@@ -1,17 +1,0 @@
-import { IsNotEmpty, IsOptional, IsString } from 'class-validator';
-
-export class ReorderFolderDto {
-  @IsString()
-  @IsNotEmpty()
-  targetId: string;
-
-  @IsOptional()
-  @IsString()
-  @IsNotEmpty()
-  afterId?: string;
-
-  @IsOptional()
-  @IsString()
-  @IsNotEmpty()
-  beforeId?: string;
-}
