--- src/folders/dtos/update-folder.dto.ts
@@ -1,8 +1,0 @@
-import { IsNotEmpty, IsOptional, IsString } from 'class-validator';
-
-export class UpdateFolderDto {
-  @IsOptional()
-  @IsString()
-  @IsNotEmpty()
-  name?: string;
-}
