+++ src/users/application/usecases/update-user-settings.usecase.ts
@@ -2,7 +2,7 @@ import { Inject, Injectable } from '@nestjs/common';
 import { USERS_REPOSITORY } from '../../domain/users.repository';
 import type { UsersRepository } from '../../domain/users.repository';
 import type { UserTheme } from '../../domain/user.types';
-import { UsersError } from '../users.error';
+import { UsersError } from '../errors/users.error';
 
 export type UpdateUserSettingsInput = {
   theme?: UserTheme;
